9 add auto build on new commits to master using GitHub actions (#10)
* feat: Add support for user-defined WiFi SSID and password

This commit modifies the `generate_wpa_conf.sh` script to allow the user to provide the WiFi SSID and password as environment variables (`BR_PROJECT_WIFI_SSID` and `BR_PROJECT_WIFI_PASSWORD`). If these variables are not set, the script prompts the user for the SSID and password. This change improves the flexibility and usability of the script.

* Added basic image build on push to master and this branch

* Added execution permission to clean-build.sh

* Fixed typo

* Added SSID and Password from secrets

* Added build on pull-request